Move place explorer search bar test out of shared
