+++ server/webdriver/shared_tests/place_explorer_i18n_test.py
@@ -94,27 +94,6 @@ class PlaceI18nExplorerTestMixin():
     WebDriverWait(self.driver,
                   self.TIMEOUT_SEC).until(EC.title_contains('アメリカ合衆国 - 人口統計'))
 
-  def test_explorer_redirect_place_explorer_populates_search_bar(self):
-    """Test the redirection from explore to place explore for single place queries populates the search bar from the URL query"""
-    usa_explore_fr_locale = '/explore?hl=fr#q=United%20States%20Of%20America'
-
-    start_url = self.url_ + usa_explore_fr_locale
-    self.driver.get(start_url)
-
-    # Assert 200 HTTP code: successful page load.
-    self.assertEqual(shared.safe_url_open(self.driver.current_url), 200)
-
-    # Wait for redirect and page load
-    redirect_finished = EC.url_changes(start_url)
-    WebDriverWait(self.driver, self.TIMEOUT_SEC).until(redirect_finished)
-    shared.wait_for_loading(self.driver)
-
-    # Ensure the query string is set in the NL Search Bar.
-    self.assertEqual(
-        find_elem(self.driver, by=By.ID,
-                  value='query-search-input').get_attribute('value'),
-        'United States Of America')
-
   def test_dev_place_page_loads_with_locale(self):
     """Ensure experimental dev place page content loads data for a continent."""
     self.driver.get(self.url_ + '/place/africa?hl=fr')
